Make cli.py default hearing_impaired value None instead of False
If the --hearing-impaired flag is NOT specified on the command line,
then args.hearing_impaired is parsed to False, and args.hearing_impaired
is passed to download_best_subtitles() overriding its own default value
of None. This commit changes cli.py to pass None instead of False when
the --hearing-impaired flag is not given on the command line.

Note that there is now no way to specify from the command line that
you do NOT want hearing-impaired subtitles. If that is necessary for
someone, I suggest adding a --no-hearing-impaired option and the
appropriate logic to pass False to download_best_subtitles()