enable podnapisi to search for series name with strings in (paren) stripped
Sometimes a series will have the year or country code in parentheses as
part of the name, but this naming can vary. In initial subtitle query be
more permissive by stripping the (data). Since it is only stripped for
the query and not subsequent matching, not a big risk since scoring can
be more restrictive and lower the score if the (data) does not match.